requirements.txt
flask~=3.1.0
grpcio~=1.74.0
grpcio-reflection~=1.74.0
web3~=7.8.0
protobuf>=6